# Fixing some performance bottlenecks