--- a/charset_normalizer/utils.py
+++ b/charset_normalizer/utils.py
@@ -204,9 +204,10 @@ def is_unprintable(character: str) -> bool:
         character.isspace() is False  # includes \n \t \r \v
         and character.isprintable() is False
         and character != "\x1A"  # Why? Its the ASCII substitute character.
-        ):
+    ):
         return True
     return False
+
 
 def any_specified_encoding(sequence: bytes, search_zone: int = 4096) -> Optional[str]:
     """