## pprofile tests

***test.py***
```python3
from glob import glob
from os.path import isdir
from charset_normalizer import detect

def performance_compare(size_coeff):
    if not isdir("./char-dataset"):
        print("This script require https://github.com/Ousret/char-dataset to be cloned on package root directory")
        exit(1)
    for tbt_path in sorted(glob("./char-dataset/**/*.*")):
        with open(tbt_path, "rb") as fp:
            content = fp.read() * size_coeff            
        detect(content)

if __name__ == "__main__":
    performance_compare(1)
```
---

### Before
```python3
pprofile --format callgrind --out cachegrind.out.original.test test.py
```
Time: 838.97 s.
[cachegrind.out.original.zip](https://github.com/Ousret/charset_normalizer/files/8598533/cachegrind.out.original.zip)
![cachegrind out original test](https://user-images.githubusercontent.com/65482418/166406184-0c5c203b-c289-4cb8-a418-704545ac46cb.jpg)


---
### After commits
```python3
pprofile --format callgrind --out cachegrind.out.commits.test test.py
```
Time: 698.53 s.
[cachegrind.out.commits.zip](https://github.com/Ousret/charset_normalizer/files/8598511/cachegrind.out.commits.zip)
![cachegrind out commits](https://user-images.githubusercontent.com/65482418/166406197-27a2aea7-ef7d-4664-8cd4-30d4bfb7e3bf.jpg)