@deedy5 2022-04-30
### line_profiler test before changes:
```
kernprof -lv performance.py
```
```
Total time: 3.58387 s
File: /charset_normalizer-master/charset_normalizer/md.py
Function: feed at line 140

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   140                                               @profile
   141                                               def feed(self, character: str) -> None:
   142   4654032    1089618.0      0.2     30.4          if (
   143   2494830     899525.0      0.4     25.1              character.isspace() is False  # includes \n \t \r \v
   144   2144442     721900.0      0.3     20.1              and character.isprintable() is False
   145     14760       4099.0      0.3      0.1              and character != "\x1A"  # Why? Its the ASCII substitute character.
   146                                                   ):
   147     14585       4963.0      0.3      0.1              self._unprintable_count += 1
   148   2494830     863770.0      0.3     24.1          self._character_count += 1

```
---
### line_profiler test after changes:
```
kernprof -lv performance.py
```
```
Total time: 1.62655 s
File: /charset_normalizer-enhanced/charset_normalizer/md.py
Function: feed at line 141

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   141                                               @profile
   142                                               def feed(self, character: str) -> None:
   143   2494830     821098.0      0.3     50.5          if is_unprintable(character):
   144     14585       4548.0      0.3      0.3              self._unprintable_count += 1
   145   2494830     800900.0      0.3     49.2          self._character_count += 1

Total time: 0.008891 s
File: /charset_normalizer-enhanced/charset_normalizer/utils.py
Function: is_unprintable at line 201

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   201                                           @lru_cache(maxsize=UTF8_MAXIMAL_ALLOCATION)
   202                                           @profile
   203                                           def is_unprintable(character: str) -> bool:
   204     14995       3686.0      0.2     41.5      if (
   205      7354       2635.0      0.4     29.6          character.isspace() is False  # includes \n \t \r \v
   206      7338       2419.0      0.3     27.2          and character.isprintable() is False
   207       303         84.0      0.3      0.9          and character != "\x1A"  # Why? Its the ASCII substitute character.
   208                                                   ):
   209       302         67.0      0.2      0.8          return True
```