@deedy5 2022-05-01
### Before commits and mypyc compilation
```
--> Charset-Normalizer Conclusions
   --> Avg: 0.0239454577s
   --> 99th: 0.242043105s
   --> 95th: 0.130321793s
   --> 50th: 0.009998223s
```
### After commits and mypyc compilation
```
--> Charset-Normalizer Conclusions
   --> Avg: 0.010750019589361701s
   --> 99th: 0.112227622s
   --> 95th: 0.054715067s
   --> 50th: 0.00445089s
```