@deedy5 2022-05-01
### Tests `/bin/performance.py` without dir char-dataset/None

---
### 1) size-increase = 1
```
python3 performance.py -s 1
```
Before
```python3
--> Charset-Normalizer Conclusions
   --> Avg: 0.03203997867803838s
   --> 99th: 0.34011s
   --> 95th: 0.16578s
   --> 50th: 0.01472s
```
After commits and mypyc compilation
```python3
--> Charset-Normalizer Conclusions
   --> Avg: 0.01026181236673774s
   --> 99th: 0.10958s
   --> 95th: 0.05292s
   --> 50th: 0.00413s
```
---
### 2) size-increase = 10
```
python3 performance.py -s 10
```
Before

```python3
--> Charset-Normalizer Conclusions
   --> Avg: 0.02120435064935065s
   --> 99th: 0.42033s
   --> 95th: 0.08811s
   --> 50th: 0.00393s
```
After commits and mypyc compilation
```python3
--> Charset-Normalizer Conclusions
   --> Avg: 0.0105862987012987s
   --> 99th: 0.13491s
   --> 95th: 0.03586s
   --> 50th: 0.00137s
```
---
### 3) size-increase = 100
```
python3 performance.py -s 100
```
Before

```python3
--> Charset-Normalizer Conclusions
   --> Avg: 0.05665621212121212s
   --> 99th: 1.70197s
   --> 95th: 0.14022s
   --> 50th: 0.00617s
```
After commits and mypyc compilation
```python3
--> Charset-Normalizer Conclusions
   --> Avg: 0.04496573593073593s
   --> 99th: 1.64798s
   --> 95th: 0.1104s
   --> 50th: 0.0037s
```