@deedy5 2022-05-01
I also made a heat map with vprof.

***heatmap_test.py***
```python3
from glob import glob
from os.path import isdir
from charset_normalizer import detect

def get_data():
    if not isdir("./char-dataset"):
        print("This script require https://github.com/Ousret/char-dataset to be cloned on package root directory")
        exit(1)
    data = []
    for tbt_path in sorted(glob("./char-dataset/**/*.*")):
        with open(tbt_path, "rb") as fp:
            content = fp.read()
            data.append(content)
    return data

def performance_compare(size_coeff):
    data = get_data()
    len_data = len(data)
    for i,d in enumerate(data, start=1):
        d = d[:750]
        detect(d)
        print(f"{i}/{len_data}")

if __name__ == "__main__":
    performance_compare(1)
```
***vprof test***
```
vprof -c h heatmap_test.py
```
***result***
[vprof_commited.zip](https://github.com/Ousret/charset_normalizer/files/8599370/vprof_commited.zip)

There are no more bottlenecks. Speeding up is now a matter of rewriting code in cython.