@deedy5 2022-05-02
I did my tests and filled in the table, the performance increased by 20% with size_coeff = 1
[ptest.zip](https://github.com/Ousret/charset_normalizer/files/8602021/ptest.zip)
[difference.xlsx](https://github.com/Ousret/charset_normalizer/files/8602022/difference.xlsx)