@deedy5 2022-05-02
***Performance comparison table***

![screen](https://user-images.githubusercontent.com/65482418/166236285-6d374283-ea9a-4587-a9ed-6b776a2be69c.jpg)
[performance_comparison.xlsx](https://github.com/Ousret/charset_normalizer/files/8602905/performance_comparison.xlsx)