@Ousret 2022-05-03
Everything looks good. Let's proceed/move on.

A bit of caution with the expressed delays by GHA. We never know what VM is behind or how scaled.
Consider using factor between chardet and charset-normalizer.