@deedy5 2022-05-03
> Everything looks good. Let's proceed/move on.
> 
> A bit of caution with the expressed delays by GHA. We never know what VM is behind or how scaled. Consider using factor between chardet and charset-normalizer.
### Performance comparison
***181*** : https://github.com/Ousret/charset_normalizer/pull/181 / [performance test](https://github.com/Ousret/charset_normalizer/runs/6211035114?check_suite_focus=true)
***183*** : https://github.com/Ousret/charset_normalizer/pull/183 / [performance test](https://github.com/Ousret/charset_normalizer/runs/6277552469?check_suite_focus=true)
***Summary***
![performance_comparison](https://user-images.githubusercontent.com/65482418/166556893-d3dd06f9-70b1-4c99-9d74-a73a0331e049.jpg)
[performance_comparison.xlsx](https://github.com/Ousret/charset_normalizer/files/8613938/performance_comparison.xlsx)